MessageGrp::sum(size_t*) -> MessageGrp::sum(unsigned long*) , resolves #45

+++ src/lib/util/group/message.h
@@ -499,7 +499,7 @@ class MessageGrp: public DescribedClass {
     virtual void sum(unsigned int* data, int n, unsigned int* = 0, int target = -1);
     virtual void sum(int* data, int n, int* = 0, int target = -1);
     virtual void sum(long* data, int n, long* = 0, int target = -1);
-    virtual void sum(size_t* data, int n, size_t* = 0, int target = -1);
+    virtual void sum(unsigned long* data, int n, unsigned long* = 0, int target = -1);
     virtual void sum(char* data, int n, char* = 0, int target = -1);
     virtual void sum(unsigned char* data, int n,
                      unsigned char* = 0, int target = -1);
