+++ src/lib/util/group/reduce.cc
@@ -157,7 +157,7 @@ MessageGrp::sum(long* data, int n, long* tmp, int target)
 }
 
 void
-MessageGrp::sum(size_t* data, int n, size_t* tmp, int target)
+MessageGrp::sum(unsigned long* data, int n, unsigned long* tmp, int target)
 {
   do_sum(this, data, n, tmp, target);
 }
